requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
idna==2.10
pycrypto==2.6.1
pycryptodome==3.19.0
pydantic==1.7.3
requests==2.25.1
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
